Revert "Add CoordinatorDelegateQueuable protocol used by view models to handle coordinator delegation in an operation queue."
This reverts commit 01bc067c2279511db3401bc419166113c31bed7a.
